% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_shader.R
\name{build_program}
\alias{build_program}
\title{Build a shader program}
\usage{
build_program(
  vert = NULL,
  tesc = NULL,
  tese = NULL,
  geom = NULL,
  frag = NULL,
  comp = NULL
)
}
\arguments{
\item{vert}{Vertex shader.}

\item{tesc}{Tessellation control shader.}

\item{tese}{Tessellation evaluation shader.}

\item{geom}{Geometry shader.}

\item{frag}{Fragment shader.}

\item{comp}{Compute shader.}
}
\value{
A non-zero value corresponding to the shader program reference.
}
\description{
This function creates a program object that represents fully processed
executable code in the OpenGL Shading Language for one or more shader
stages.

All arguments expect the source code of each shader. The source code can be
provided as a character vector, or as a path to the source code file. At
least the source code of one shader must be supplied. A typical use case is
to supply the source code of the vertex and fragment shaders.
}
