% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GL}
\alias{GL}
\title{OpenGL integer preprocessor constants}
\format{
An object of class \code{list} of length 5689.
}
\usage{
GL
}
\description{
\code{GL} is a list of OpenGL integer constants.
}
\keyword{datasets}
