% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glfw-package.R
\docType{package}
\name{glfw-package}
\alias{glfw}
\alias{glfw-package}
\title{glfw: R Interface to the GLFW C Library}
\description{
GLFW is a lightweight utility library for use with OpenGL: \url{https://www.glfw.org/}. It provides programmers with the ability to create and manage windows and OpenGL contexts, as well as handle joystick, keyboard and mouse input. This package provides an R interface to the underlying C library implementation.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.magno@gmail.com} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

}
\keyword{internal}
