% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GLFW}
\alias{GLFW}
\title{GLFW integer preprocessor constants}
\format{
An object of class \code{list} of length 292.
}
\usage{
GLFW
}
\description{
\code{GL} is a list of GLFW integer constants.
}
\keyword{datasets}
