% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_types.R
\docType{data}
\name{uniform_types}
\alias{uniform_types}
\title{Mapping of GL constants to GLSL uniform types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 105 rows and 2 columns.
}
\usage{
uniform_types
}
\description{
\code{uniform_types} is a data frame of GLFW integer constants to the
corresponding GLSL uniform types.
}
\keyword{datasets}
